This is a major commit.[working version]
this is the working copy of primeNumber checking javascript file. this
snippet will write the output data to the 'PrimeNumber.txt' file also.

this version needed to have some improvements. they are commented in the
first section of the file, after the copyright notification.
I'd tried alot to implement using 'node's readline module. but it
seems not working till

Signed-off-by: Joice Joseph <josukutty@outlook.com>